pyproject.toml
[project]
name = "email-dagster-dbt-bigquery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "pandas>=2.3.3",
    "pandas-gbq>=0.33.0",
    "tqdm>=4.67.1",
]
